requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.38
aiohttp==3.13.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-ollama==0.2.3
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
openai==1.64.0
python-jose==3.4.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
tiktoken==0.9.0
chromadb==0.6.3
